Handle edge-cases better, add isEqual, fix tests
Should handle being passed null/undefined in all functions, and will terminate faster when given an empty set/array.

Added a function to test whether 2 sets are equal, as using === does not work.

Fixed some tests with typos that should have failed, but passed because expect.js reports sets to be equal if they are the same size, even if they are different.